+++ src/Microsoft.SqlTools.Hosting/Hosting/Contracts/AdminServicesProviderOptions.cs
@@ -10,7 +10,11 @@ namespace Microsoft.SqlTools.Hosting.Contracts
     /// </summary>
     public class AdminServicesProviderOptions
     {
-        public ServiceOption[] Options { get; set; }
+        public ServiceOption[] DatabaseInfoOptions { get; set; }
+
+        public ServiceOption[] DatabaseFileInfoOptions { get; set; }
+
+        public ServiceOption[] FileGroupInfoOptions { get; set; }
     }
 }
