+++ src/Microsoft.SqlTools.Hosting/Hosting/Contracts/DmpServerCapabilities.cs
@@ -17,5 +17,7 @@ namespace Microsoft.SqlTools.Hosting.Contracts
         public string ProviderDisplayName { get; set; }
 
         public ConnectionProviderOptions ConnectionProvider { get; set; }
+
+        public AdminServicesProviderOptions AdminServicesProviderOptions { get; set; }
     }
 }
