+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/AzureSqlDbHelper.cs
