+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/CUtils.cs
