+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/DataContainer.cs
