+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/DatabasePrototype.cs
