+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/DatabasePrototype100.cs
