+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/DatabasePrototype110.cs
