+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/DatabasePrototype80.cs
