+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/DatabasePrototype80SP3.cs
