+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/DatabasePrototype90.cs
