+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/DatabasePrototypeAzure.cs
