+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/DatabaseTaskHelper.cs
