+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/ICloneable.cs
