+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/IManagedConnection.cs
