+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/ISandboxLoader.cs
