+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/SharedConnectionUtil.cs
