+++ src/Microsoft.SqlTools.ServiceLayer/Admin/Common/Utils.cs
