+++ src/Microsoft.SqlTools.ServiceLayer/Connection/ConnectionService.cs
