+++ src/Microsoft.SqlTools.ServiceLayer/project.json
@@ -21,7 +21,7 @@
     "Newtonsoft.Json": "9.0.1",
     "System.Data.Common": "4.1.0",
     "System.Data.SqlClient": "4.4.0-sqltools-24613-04",
-    "Microsoft.SqlServer.Smo": "140.17049.0",
+    "Microsoft.SqlServer.Smo": "140.17050.0",
     "Microsoft.SqlServer.Management.SqlScriptPublishModel": "140.17049.0",
     "System.Security.SecureString": "4.0.0",
     "System.Collections.Specialized": "4.0.1",
