+++ test/Microsoft.SqlTools.ServiceLayer.Test.Common/project.json
@@ -18,7 +18,7 @@
     "System.Runtime.Serialization.Primitives": "4.1.1",
     "System.Data.Common": "4.1.0",
     "System.Data.SqlClient": "4.4.0-sqltools-24613-04",
-    "Microsoft.SqlServer.Smo": "140.17049.0",
+    "Microsoft.SqlServer.Smo": "140.17050.0",
     "System.Security.SecureString": "4.0.0",
     "System.Collections.Specialized": "4.0.1",
     "System.ComponentModel.TypeConverter": "4.1.0",
