--- test/Microsoft.SqlTools.ServiceLayer.UnitTests/Admin/CreateLoginTests.cs
@@ -1,16 +1,0 @@
-﻿//
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-//
-
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Microsoft.SqlTools.ServiceLayer.UnitTests.Admin
-{
-    public class CreateLoginTests
-    {
-    }
-}
